+++ codegen/java/src/main/kotlin/io/spine/protodata/codegen/java/JavaValueConverter.kt
@@ -33,15 +33,13 @@ import io.spine.protodata.Type.KindCase.PRIMITIVE
 import io.spine.protodata.Value
 import io.spine.protodata.type.ValueConverter
 import io.spine.tools.code.Java
-import io.spine.tools.code.Language
-import io.spine.tools.code.SlashAsteriskCommentLang
 
 /**
  * A [ValueConverter] which converts values into Java expressions.
  */
 @Suppress("TooManyFunctions")
 public class JavaValueConverter(
-    private val typeConverter: JavaTypeConvention
+    private val typeConverter: MessageTypeConvention
 ) : ValueConverter<Expression, Java>() {
 
     override fun toNull(type: Type): Expression = Null
