+++ codegen/java/src/main/kotlin/io/spine/protodata/codegen/java/MessageTypeConvention.kt
@@ -26,35 +26,22 @@
 
 package io.spine.protodata.codegen.java
 
-import io.kotest.matchers.shouldBe
-import io.kotest.matchers.shouldNotBe
-import io.spine.protodata.test.TypesTestEnv.enumTypeName
-import io.spine.protodata.test.TypesTestEnv.messageTypeName
-import io.spine.protodata.test.TypesTestEnv.typeSystem
-import kotlin.io.path.Path
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
+import io.spine.protodata.TypeName
+import io.spine.protodata.type.GeneratedDeclaration
+import io.spine.protodata.type.TypeSystem
+import io.spine.tools.code.Java
 
-@DisplayName("`JavaTypeConvention` should")
-class JavaTypeConventionSpec {
+/**
+ * A convention which governs the Java message class declarations.
+ */
+public class MessageTypeConvention(
+    typeSystem: TypeSystem
+) : BaseJavaTypeConvention(typeSystem) {
 
-    @Test
-    fun `convert a message type name into a Java class name`() {
-        val convention = JavaTypeConvention(typeSystem)
-        val declaration = convention.declarationFor(messageTypeName)
-        declaration shouldNotBe null
-        val (cls, path) = declaration
-        cls.binary shouldBe "ua.acme.example.Foo"
-        path shouldBe Path("ua/acme/example/Foo.java")
-    }
-
-    @Test
-    fun `convert an enum type name into a Java class name`() {
-        val convention = JavaTypeConvention(typeSystem)
-        val declaration = convention.declarationFor(enumTypeName)
-        declaration shouldNotBe null
-        val (cls, path) = declaration
-        cls.binary shouldBe "ua.acme.example.Kind"
-        path shouldBe Path("ua/acme/example/Kind.java")
+    override fun declarationFor(name: TypeName): GeneratedDeclaration<Java, ClassName> {
+        val file = typeSystem.findMessageOrEnum(name)?.second
+        check(file != null) { "Unknown type `${name.typeUrl}`." }
+        val cls = name.javaClassName(declaredIn = file)
+        return GeneratedDeclaration(cls, cls.javaFile)
     }
 }
